LayoutTests/fast/html/details-mouse-click-expected.txt

details0's open attribute is expected to be'true' and is.
details0's open attribute is expected to be'false' and is.
details1's open attribute is expected to be'true' and is.
details1's open attribute is expected to be'false' and is.
details2's open attribute is expected to be'true' but instead is 'false'.
details2's open attribute is expected to be'true' but instead is 'false'.
FAIL
